+++ downloader/symbols/database.py
@@ -22,12 +22,13 @@ class database():
         self.table = self.dataset.table('symbols', schema)
 
 
+    def delete(self):
+        if self.table.exists():
+            self.table.delete()
+
+
     def create(self):
         self.table.create()
-
-
-    def delete(self):
-        self.table.delete()
 
 
     def upload(self, filename):
